+++ NINA.Plugin.TargetScheduler/Controls/Reporting/ReportingManagerViewVM.cs
@@ -116,6 +116,7 @@ namespace NINA.Plugin.TargetScheduler.Controls.Reporting {
                 targets = context.TargetSet.AsNoTracking().Where(t => t.ProjectId == selectedProjectId).ToList();
             }
 
+            targets = targets.OrderBy(t => t.Name).ToList();
             targets.ForEach(t => {
                 choices.Add(new KeyValuePair<int, string>(t.Id, t.Name));
             });