+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Target Scheduler
 
-## 5.1.8.0 (beta) - 2025-03-XX
+## 5.1.8.0 (beta) - 2025-03-22
 * Added ability to manually grade all pending exposure plans for a target.
 * Targets in Reporting dropdown are now sorted by name.
 * Added Exposure Template name to Acquired Images row detail view.
-* Other misc issues.
+* Added database busy timeout to avoid locked errors.
+* Other misc fixes.
 
 ## 5.1.7.0 (beta) - 2025-03-18
 * Smart exposure selector will now select the exposure with lowest percent complete when multiple plans have equal highest avoidance scores.